refactored util file
